Refactor imports and update file paths in main.py
